refactor(parser): expose specification builder methods
