+++ lib/index.js
@@ -2,6 +2,11 @@
 
 const getSpecificationObject = require('./helpers/getSpecificationObject');
 
+const createSpecification = require('./helpers/createSpecification');
+const parseApiFileContent = require('./helpers/parseApiFileContent');
+const updateSpecificationObject = require('./helpers/updateSpecificationObject');
+const finalizeSpecificationObject = require('./helpers/finalizeSpecificationObject');
+
 /**
  * Generates the specification.
  * @function
@@ -32,3 +37,8 @@ module.exports = options => {
     throw new Error(err);
   }
 };
+
+module.exports.createSpecification = createSpecification;
+module.exports.parseApiFileContent = parseApiFileContent;
+module.exports.updateSpecificationObject = updateSpecificationObject;
+module.exports.finalizeSpecificationObject = finalizeSpecificationObject;
